+++ src/main/java/modelo/ProductoDAO.java
@@ -25,7 +25,7 @@ public class ProductoDAO {
             ps = con.prepareStatement(sql);
             rs = ps.executeQuery();
             while (rs.next()) {
-                listaProductos.add(new Producto(rs.getInt(1), rs.getString(2), rs.getDouble(3), rs.getInt(4), rs.getByte(5)));
+                listaProductos.add(new Producto(rs.getInt(1), rs.getString(2), rs.getBigDecimal(3), rs.getInt(4), rs.getByte(5)));
             }
             logger.info("Se logró listar los productos exitósamente");
         } catch (SQLException e) {
@@ -42,7 +42,7 @@ public class ProductoDAO {
             con = conexion.conectar();
             ps = con.prepareStatement(sql);
             ps.setString(1, oProducto.getNombre());
-            ps.setDouble(2, oProducto.getPrecio());
+            ps.setBigDecimal(2, oProducto.getPrecio());
             ps.setInt(3, oProducto.getStock());
             ps.setByte(4, oProducto.getEstado());
             ps.executeUpdate();
@@ -63,7 +63,7 @@ public class ProductoDAO {
             ps.setInt(1, id);
             rs = ps.executeQuery();
             while (rs.next()) {                
-                oProducto = new Producto(rs.getInt(1), rs.getString(2), rs.getDouble(3), rs.getInt(4), rs.getByte(5));
+                oProducto = new Producto(rs.getInt(1), rs.getString(2), rs.getBigDecimal(3), rs.getInt(4), rs.getByte(5));
             }
             logger.info("Se logró encontrar el producto exitósamente");
         } catch (SQLException e) {
@@ -80,7 +80,7 @@ public class ProductoDAO {
             con = conexion.conectar();
             ps = con.prepareStatement(sql);
             ps.setString(1, nuevoProducto.getNombre());
-            ps.setDouble(2, nuevoProducto.getPrecio());
+            ps.setBigDecimal(2,nuevoProducto.getPrecio());
             ps.setInt(3, nuevoProducto.getStock());
             ps.setByte(4, nuevoProducto.getEstado());
             ps.setInt(5, nuevoProducto.getId_producto());
